fix : jpa 데이터 타입 관련 매핑 코드 수정

+++ module-auth/src/main/kotlin/com/raillylinker/module_auth/jpa_beans/db1_main/entities/Db1_RaillyLinkerCompany_TotalAuthJoinTheMembershipWithOauth2Verification.kt
@@ -16,7 +16,7 @@ import java.time.LocalDateTime
 class Db1_RaillyLinkerCompany_TotalAuthJoinTheMembershipWithOauth2Verification(
     @Column(name = "oauth2_type_code", nullable = false, columnDefinition = "TINYINT UNSIGNED")
     @Comment("oauth2 종류 (1 : GOOGLE, 2 : NAVER, 3 : KAKAO)")
-    var oauth2TypeCode: Byte,
+    var oauth2TypeCode: Short,
 
     @Column(name = "oauth2_id", nullable = false, columnDefinition = "VARCHAR(50)")
     @Comment("OAuth2 로그인으로 얻어온 고유값")
